subscriber: add ability to disable ANSI without crate feature
Currently it is not possible to disable ANSI in `fmt::Subscriber`
without enabling the `ansi` crate feature. This makes it difficult for
users to implement interoperable settings that are controllable with
crate features without having to pull in the dependencies `ansi` does.

This adds a simple `fmt::Subscriber::disable_ansi()` that is available
even without the `ansi` crate feature.